Light Baked

+++ Assets/_Main/Scripts/Upgrades/BaseUpgradeSo.cs
@@ -5,8 +5,8 @@ namespace _Main.Scripts.Upgrades
     [CreateAssetMenu(fileName = "BaseUpgradeSO", menuName = "Upgrades/Base", order = 0)]
     public class BaseUpgradeSo : ScriptableObject
     {
-        [field:SerializeField] public int MaxValue { get; private set; }
-        [field:SerializeField] public int MinValue { get; private set; }
+        [field:SerializeField] public float MaxValue { get; private set; }
+        [field:SerializeField] public float MinValue { get; private set; }
         [field:SerializeField, Multiline]public string Description { get; private set; }
         [field:SerializeField]public Sprite Image { get; private set; }
     }
